Remove a couple of the arithmetic exception throws because it was hitting on things that it shouldn't (e.g., encoding the bits in a double).
Make sure the term stats can handle the new long getTotalOccurrences.

+++ src/com/sun/labs/minion/retrieval/WeightingComponents.java
@@ -106,7 +106,7 @@ public class WeightingComponents {
      * The total number of occurrences of term <em>t</em> in the whole
      * collection.  Document-level statistic.
      */
-    public int Ft;
+    public long Ft;
 
     /**
      * The total number of documents containing term <em>t</em>.
